+++ CHANGELOG.md
@@ -4,6 +4,16 @@ The changelog can be found at https://github.com/mattermost/mattermost-plugin-gi
 
 ## 0.2.0 - unreleased
 
+- Finish private repositories on/off [#18](https://github.com/manland/mattermost-plugin-gitlab/pull/18)
+- Finish all webhook implementation [#16](https://github.com/manland/mattermost-plugin-gitlab/pull/16): 
+    - MergeEvent
+    - IssueEvent
+    - IssueCommentEvent
+    - MergeCommentEvent
+    - PushEvent
+    - PipelineEvent
+    - TagEvent
+
 ## 0.1.0 - 2019-04-17
 
 - Get all code from [mattermost-plugin-github](https://github.com/mattermost/mattermost-plugin-github/) and adapt it for gitlab